@@ -15,7 +15,7 @@
     <name>cepsearch</name>
     <description>A API to consumes BrasilAPI CEP</description>
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
     </properties>
     <dependencies>
         <dependency>
