@@ -3,8 +3,8 @@ package com.cepsearch.annotations.impl;
 import com.cepsearch.annotations.PostalCodeValidation;
 import org.apache.commons.lang3.StringUtils;
 
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
 
 public class PostalCodeValidationImpl implements ConstraintValidator<PostalCodeValidation, String> {
 
